requirements.txt
# 核心依赖
futu-api>=6.0.0
akshare>=1.9.0
tushare>=1.2.0
pandas>=1.5.0
numpy>=1.21.0

# Web界面
flask>=2.0.0

# 其他工具
python-dateutil>=2.8.0
pytz>=2021.3